Move the XML-RPC and admin-locator helpers into config/
The XML-RPC request builder was copy-pasted verbatim into three specs, so a
wire-format change meant three edits in lockstep; it now lives in
config/xmlrpc.ts alongside expectXmlRpcFault(), which only one of the three
had. formTableRow() and userRowLocator() were each duplicated across two
specs and move to config/admin.ts.

admin.ts also loses rowActionLocator(), which nothing called and which
pairs a #post-<id> base with row actions the suite only ever reads off
#user-<id> rows. editPhp() and postNewPhp() are now used at the call sites
that were hand-writing the same URLs, and noticeLocator() stays as an
internal of noticeWith().

Docblocks in the touched specs state their invariant in the present tense
instead of recounting the 0.5.5 bug that prompted them.
