+++ tests/e2e/specs/admin/dashboard-console.spec.ts
@@ -3,14 +3,14 @@
  * `dwpb.commentsSupported` is `false` — unreachable via the plugin's normal
  * default state, so exercised here via a test fixture toggle.
  *
- * Regression guard (fixed since v0.5.5): `disable-blog-admin.js`'s Dashboard
- * case used to call `.parentNode` on `document.querySelector(
- * '.welcome-icon.welcome-comments' )`, which is `null` on WordPress 6.1+ —
- * an uncaught `TypeError` inside the `DOMContentLoaded` handler that killed
- * every later `case` in that handler for the rest of the page load. An
- * uncaught exception surfaces to Playwright as a `pageerror` event, not
- * `console` — the assertion below binds to `pageerror` specifically; the
- * `console` collector is kept only to enrich the failure message.
+ * Regression guard: `document.querySelector( '.welcome-icon.welcome-comments'
+ * )` is `null` on WordPress 6.1+, so `disable-blog-admin.js`'s Dashboard case
+ * must not call `.parentNode` on it directly -- doing so throws an uncaught
+ * `TypeError` inside the `DOMContentLoaded` handler, killing every later
+ * `case` in that handler for the rest of the page load. An uncaught exception
+ * surfaces to Playwright as a `pageerror` event, not `console` — the
+ * assertion below binds to `pageerror` specifically; the `console` collector
+ * is kept only to enrich the failure message.
  *
  * This overlaps with `filters/comments-unsupported.spec.ts`'s final test
  * (also asserts a clean dashboard in this state) but is kept separate: that
@@ -24,25 +24,10 @@
 import { test, expect } from '@wordpress/e2e-test-utils-playwright';
 
 /**
- * External dependencies
- */
-import type { Page } from '@playwright/test';
-
-/**
  * Internal dependencies
  */
+import { formTableRow } from '../../config/admin';
 import { setFixtures, resetFixtures, FIXTURE_TOGGLES } from '../../config/fixtures';
-
-/**
- * Row-of-a-`.form-table` locator, matched by the `<label for="...">` inside
- * it. Mirrors `formTableRow()` in `settings-screens.spec.ts`, kept local here.
- *
- * @param page     Page under test.
- * @param labelFor The `for` attribute of a `<label>` inside the target row.
- */
-function formTableRow( page: Page, labelFor: string ) {
-	return page.locator( 'tr', { has: page.locator( `label[for="${ labelFor }"]` ) } );
-}
 
 test.describe( 'admin: dashboard console (commentsSupported === false)', () => {
 	test.beforeAll( async ( { requestUtils } ) => {
